requirements.txt
fastapi
uvicorn
jinja2
aiofiles
google-cloud-firestore
requests
python-multipart
firebase-admin
sentence-transformers
torch
openai
dateparser